Add optional type.
Optional type is a new kind of builtin type allowing to declare values that
could have a value but could also don't have a value.

We declare an optional value like this :

        x is optional integer

Now, if we do :

        x.is_set()

The expression 'x.is_set()' is false since we never assigned the variable 'x'.

Now, to assign a value to x, one could simply do :

        x.set(32)

To get the value of x, just do :

        x.get()

Be careful, is x.is_set() is false, calling 'get' will generate a runtime
program exit.

We introduce with this commit two samples :
- test-optional.ez is a simple program using optional
- arbre.ez is a more complex program showing how to use optional values to
  do binary trees.

Enjoy ;)